requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
cryptography==46.0.3
pycparser==3.0
typing_extensions==4.15.0
